# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: GlowCycle - Smart Period Tracker & Women's Health Companion API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users/profile:
    put:
      operationId: updateProfile
      tags: [users]
      summary: Update user profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProfileBody"
      responses:
        "200":
          description: Profile updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
...
  /products:
    get:
      operationId: getProducts
      tags: [products]
      summary: Get all store products
      parameters:
        - name: category
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"
    post:
      operationId: createProduct
      tags: [products]
      summary: Admin - Create a product
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProductBody"
      responses:
        "201":
          description: Product created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
  /products/{productId}:
    put:
      operationId: updateProduct
      tags: [products]
      summary: Admin - Update a product
      parameters:
        - name: productId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProductBody"
      responses:
        "200":
          description: Product updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
    delete:
      operationId: deleteProduct
      tags: [products]
      summary: Admin - Delete a product
      parameters:
        - name: productId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
...
components:
  schemas:
...
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required: [message]
...
    User:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        age:
          type: integer
        heightCm:
          type: number
        weightKg:
          type: number
        bmi:
          type: number
        bmiCategory:
          type: string
        isAdmin:
          type: boolean
        createdAt:
          type: string
      required: [id, name, email, age, heightCm, weightKg, bmi, bmiCategory, isAdmin, createdAt]
    UpdateProfileBody:
      type: object
      properties:
        name:
          type: string
        age:
          type: integer
        heightCm:
          type: number
        weightKg:
          type: number
...
    Product:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
        price:
          type: number
        category:
          type: string
        imageUrl:
          type: string
        inStock:
          type: boolean
        createdAt:
          type: string
      required: [id, name, description, price, category, inStock, createdAt]
    CreateProductBody:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        price:
          type: number
        category:
          type: string
        imageUrl:
          type: string
        inStock:
          type: boolean
      required: [name, description, price, category, inStock]
...