# lib/api-zod/src/generated/types/createProductBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * GlowCycle - Smart Period Tracker & Women's Health Companion API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateProductBody {
  name: string;
  description: string;
  price: number;
  category: string;
  imageUrl?: string;
  inStock: boolean;
}